Move test to its own module
